+++ pkg/engine/destroy.go
@@ -61,7 +61,7 @@ func newDestroySource(
 	}
 
 	// Like Update, if we're missing plugins, attempt to download the missing plugins.
-	if err := ensurePluginsAreInstalled(client, plugins); err != nil {
+	if err := ensurePluginsAreInstalled(plugins); err != nil {
 		logging.V(7).Infof("newDestroySource(): failed to install missing plugins: %v", err)
 	}